+++ src/main/java/org/example/sqi_images/auth/authentication/AuthenticationInterceptor.java
@@ -1,4 +1,4 @@
-package org.example.sqi_images.common.authentication;
+package org.example.sqi_images.auth.authentication;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
